+++ CHANGELOG.md
@@ -28,5 +28,7 @@ and this project adheres to
 - Refactor the LRS client to be asynchronous
 - Fix count of 0 in all video endpoints
 - Require Python minimum version of 3.8
+- Encapsulate statements pre-processing in a Mixin class
+- Factorize Video indicators
 
 [unreleased]: https://github.com/openfun/warren